- VeSTige header struct VstTimeInfo touched.
git-svn-id: http://svn.code.sf.net/p/qtractor/code_attic/trunk@2833 304e7ffa-9037-4501-99c0-529757c9bbcc

+++ src/vestige/aeffectx.h
@@ -297,8 +297,8 @@ typedef struct VstTimeInfo
     double barStartPos;
     double cycleStartPos;
     double cycleEndPos;
-    double timeSigNumerator;
-    double timeSigDenominator;
+    int32_t   timeSigNumerator;
+    int32_t   timeSigDenominator;
     int32_t   smpteOffset;
     int32_t   smpteFrameRate;
     int32_t   samplesToNextClock;
